[skyapi] [firmware] [ci] refs #287 - Refactor entropy pool
- Accumulkated entopy buffer and mixing functions moved to skycoin-api
- Device-specific salt sources in formware code
- Use random_salted_buffer for ECDSA K random nonces
