+++ skycoin-api/tools/rand.h
@@ -32,18 +32,4 @@ uint32_t random_uniform(uint32_t n);
 void random_buffer(uint8_t* buf, size_t len);
 void random_permute(char* buf, size_t len);
 
-#ifndef RAND32_FUNC
-  #define RAND32_FUNC random32
-#endif
-
-#ifndef RAND_FUNC
-  #define RAND_FUNC random_buffer
-#endif
-
-#define STRING2(x) #x
-#define STRING(x) STRING2(x)
-
-#pragma message (__FILE__ "[" STRING(__LINE__) "]: Random 32-bits function " STRING(RAND32_FUNC))
-#pragma message (__FILE__ "[" STRING(__LINE__) "]: Random buffer function " STRING(RAND_FUNC))
-
 #endif
