+++ tiny-firmware/bootloader/Makefile
@@ -19,9 +19,12 @@ OBJS += ../gen/fonts.o
 
 #skycoin-crypto
 OBJS += $(TOP_DIR)vendor/skycoin-crypto/skycoin_signature.o
+OBJS += $(TOP_DIR)vendor/skycoin-crypto/skycoin_crypto.o
 OBJS += $(TOP_DIR)vendor/skycoin-crypto/tools/curves.o
 OBJS += $(TOP_DIR)vendor/skycoin-crypto/tools/hmac.o
 OBJS += $(TOP_DIR)vendor/skycoin-crypto/tools/secp256k1.o
+OBJS += $(TOP_DIR)vendor/skycoin-crypto/tools/rand.o
+OBJS += $(TOP_DIR)vendor/skycoin-crypto/tools/entropypool.o
 # OBJS += $(TOP_DIR)vendor/skycoin-crypto/tools/bip32.o
 
 #end skycoin-crypto
